+++ CHANGELOG.md
@@ -11,7 +11,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added ahead/behind info to `mepo status`
+
 ### Changed
+
+- Moved coloring of `git status` output from `src/mepo/git.py` to `src/mepo/command/status/py`
+- Streamlined output processing of `git status`
+- Migrated mepo management from rye to uv
+
+## [2.3.2] - 2025-04-24
+
+### Fixed
+
+- Fixed `checkout` command to use `--recurse-submodules` if component has submodules
+
+## [2.3.1] - 2025-04-16
+
+### Fixed
+
+- Fixed `clone` to allow for fixtures without a `develop:` key in `components.yaml`
 
 ## [2.3.0] - 2025-01-12